Move PrefsTab to gui.preftabs
