+++ src/main/java/net/sf/jabref/ExternalTab.java
@@ -31,6 +31,7 @@ import net.sf.jabref.gui.help.HelpDialog;
 
 import com.jgoodies.forms.builder.DefaultFormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
+import net.sf.jabref.gui.preftabs.PrefsTab;
 
 public class ExternalTab extends JPanel implements PrefsTab {
