+++ src/main/java/net/sf/jabref/GeneralTab.java
@@ -35,6 +35,7 @@ import net.sf.jabref.gui.help.HelpAction;
 
 import com.jgoodies.forms.builder.DefaultFormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
+import net.sf.jabref.gui.preftabs.PrefsTab;
 
 public class GeneralTab extends JPanel implements PrefsTab {
