+++ src/main/java/net/sf/jabref/NetworkTab.java
@@ -27,6 +27,7 @@ import javax.swing.event.ChangeListener;
 
 import com.jgoodies.forms.builder.DefaultFormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
+import net.sf.jabref.gui.preftabs.PrefsTab;
 
 public class NetworkTab extends JPanel implements PrefsTab {
