+++ src/main/java/net/sf/jabref/groups/GroupsPrefsTab.java
@@ -23,7 +23,7 @@ import javax.swing.*;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.PrefsTab;
+import net.sf.jabref.gui.preftabs.PrefsTab;
 
 import com.jgoodies.forms.builder.DefaultFormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
